refactor: BoardInitializer 생성
